Split phase audit truth tests
