+++ tests/docs_truth/phase_audit/phase_plan.rs
@@ -1,0 +1,210 @@
+use super::super::*;
+
+#[test]
+fn phase_plan_records_recovered_progress_and_next_slice() {
+    let plan = read("docs/PHASE_PLAN.md");
+
+    for required in [
+        "Recovery Point",
+        "183d140c",
+        "Completed Evidence",
+        "Current Phase",
+        "Next Small Slice",
+        "Sync/Async are real effects",
+        "typed allocators",
+        "actors live in std first",
+        "AST/HIR traversal is tooling/metaprogramming",
+        "type matching and behavior association are separate",
+        "JSON is compiler-owned IR output",
+        "YAML is human-authored config/spec input",
+        "build.zen is deterministic comptime build graph",
+        "Dev UX",
+        "Agent UX",
+        "MoonBit-style toolchain integration",
+        "VS Code extension",
+        "language server",
+        "agent-readable diagnostics",
+        "Machine-readable project graph",
+        "structured fix suggestions",
+        "deterministic_build_graph_creates_one_executable_target",
+        "build_graph_rejects_undeclared_host_effects",
+        "parse_project_build_zen_example",
+        "parse_shorthand_enum_variant_expr_and_pattern",
+        "parsed_project_build_zen_lowers_to_executable_and_test_graph",
+        "build_program_lowering_collects_test_target",
+        "build_program_lowering_collects_library_target",
+        "build_program_lowering_collects_target_dependencies_and_features",
+        "build_program_lowering_collects_static_block_targets",
+        "build_program_lowering_rejects_dynamic_target_adds",
+        "build_graph_orders_targets_before_dependents",
+        "build_graph_rejects_cyclic_target_dependencies",
+        "build_program_lowering_rejects_cyclic_target_dependencies",
+        "build_graph_rejects_self_target_dependencies",
+        "build_program_lowering_rejects_self_target_dependencies",
+        "emit_json_hir_outputs_checked_declaration_graph",
+        "schema_version: 0",
+        "emit_json_mir_outputs_checked_minimal_function_graph",
+        "emit_json_target_yaml_validates_minimal_target_schema",
+        "emit_json_target_yaml_validates_backend_schema",
+        "emit_json_target_yaml_validates_c_backend_flags",
+        "emit_json_target_yaml_rejects_empty_c_backend_flags",
+        "emit_json_target_yaml_rejects_layout_overrides",
+        "emit_json_target_yaml_rejects_unsupported_backend_codegen",
+        "build_graph_rejects_unknown_target_dependencies",
+        "build_program_lowering_rejects_unknown_target_dependencies",
+        "build_program_lowering_rejects_unsupported_package_targets",
+        "build_program_lowering_rejects_unsupported_link_targets",
+        "build_program_lowering_rejects_gated_package_fields",
+        "build_program_lowering_rejects_gated_link_fields",
+        "build_program_lowering_accepts_declared_env_reads",
+        "build_program_lowering_rejects_undeclared_env_reads",
+        "production_rust_files_stay_below_cleanup_threshold",
+        "zen_source_files_stay_below_cleanup_threshold",
+        "stdlib/io/mux/uring.zen",
+        "source_ast_does_not_carry_dead_char_literal_nodes",
+        "parser_type_declaration_suffixes_use_owned_keyword_enum",
+        "every tracked Rust source file",
+        "BuildGraphExecutionContext",
+        "emit_json_build_graph_outputs_project_build_graph",
+        "emit_json_build_graph_outputs_library_target",
+        "emit_json_build_graph_outputs_target_dependencies_and_features",
+        "emit_json_build_graph_rejects_cyclic_target_dependencies",
+        "emit_json_build_graph_rejects_self_target_dependencies",
+        "emit_json_build_graph_rejects_unknown_target_dependencies",
+        "emit_json_build_graph_rejects_dynamic_target_adds",
+        "emit_json_build_graph_rejects_duplicate_library_target_fields",
+        "emit_json_build_graph_rejects_missing_library_exports",
+        "emit_json_build_graph_rejects_invalid_library_exports_type",
+        "emit_json_build_graph_rejects_empty_library_exports",
+        "emit_json_build_graph_rejects_unknown_test_target_fields",
+        "emit_json_build_graph_rejects_unknown_library_target_fields",
+        "emit_json_build_graph_rejects_unsupported_package_targets",
+        "emit_json_build_graph_rejects_unsupported_link_targets",
+        "emit_json_build_graph_rejects_gated_package_fields",
+        "emit_json_build_graph_rejects_gated_link_fields",
+        "build_command_build_zen_rejects_unsupported_package_targets",
+        "build_command_build_zen_rejects_unsupported_link_targets",
+        "build_command_build_zen_rejects_dynamic_target_adds_before_execution",
+        "direct_file_command_build_zen_rejects_unsupported_package_targets",
+        "direct_file_command_build_zen_rejects_unsupported_link_targets",
+        "check_command_build_zen_rejects_unsupported_package_targets",
+        "check_command_build_zen_rejects_unsupported_link_targets",
+        "test_command_build_zen_rejects_unsupported_package_targets",
+        "test_command_build_zen_rejects_unsupported_link_targets",
+        "emit_command_build_zen_rejects_unsupported_package_targets",
+        "emit_command_build_zen_rejects_unsupported_link_targets",
+        "build_graph_command_rejects_unsupported_package_targets",
+        "build_graph_command_rejects_unsupported_link_targets",
+        "build_zen_commands_reject_package_fields",
+        "build_zen_commands_reject_link_fields",
+        "build_zen_commands_reject_dynamic_target_adds",
+        "build_zen_commands_reject_duplicate_executable_target_fields",
+        "build_zen_commands_reject_missing_required_executable_target_fields",
+        "build_zen_commands_reject_invalid_executable_target_field_types",
+        "build_zen_commands_reject_duplicate_test_target_fields",
+        "build_zen_commands_reject_missing_required_test_target_fields",
+        "build_zen_commands_reject_invalid_test_target_field_types",
+        "build_zen_commands_reject_unknown_test_target_fields",
+        "build_zen_commands_reject_unknown_library_target_fields",
+        "emit_json_build_graph_rejects_undeclared_host_effects",
+        "emit_json_build_graph_outputs_declared_env_read_effects",
+        "emit_json_build_graph_outputs_declared_file_read_effects",
+        "emit_json_build_graph_outputs_wildcard_fallback_declared_file_read_effects",
+        "emit_json_build_graph_outputs_identifier_fallback_declared_file_read_effects",
+        "emit_json_build_graph_rejects_undeclared_file_read_effects",
+        "emit_json_build_graph_rejects_undeclared_host_effects_before_test_target_lowering",
+        "emit_json_build_graph_rejects_undeclared_host_effects_before_library_target_lowering",
+        "emit_json_build_graph_rejects_undeclared_host_effects_before_target_metadata_lowering",
+        "build_graph_command_compiles_single_executable_target",
+        "build_graph_command_compiles_multiple_executable_targets",
+        "build_graph_command_compiles_executable_dependencies_first",
+        "build_graph_command_accepts_declared_env_read_with_fallback",
+        "build_graph_command_accepts_declared_env_read_for_multiple_targets",
+        "build_graph_command_accepts_wildcard_fallback_declared_env_read_for_multiple_targets",
+        "build_graph_command_accepts_identifier_fallback_declared_env_read_for_multiple_targets",
+        "build_graph_command_multi_target_rejects_env_read_without_fallback_before_execution",
+        "build_graph_command_accepts_declared_file_read_effects",
+        "build_graph_command_accepts_declared_file_read_effects_for_multiple_targets",
+        "build_graph_command_accepts_wildcard_fallback_declared_file_read_effects",
+        "build_graph_command_accepts_identifier_fallback_declared_file_read_effects",
+        "build_graph_command_accepts_wildcard_fallback_declared_file_read_effects_for_multiple_targets",
+        "build_graph_command_accepts_identifier_fallback_declared_file_read_effects_for_multiple_targets",
+        "build_graph_command_rejects_undeclared_file_read_effects_before_execution",
+        "build_graph_command_multi_target_rejects_file_read_without_fallback_before_execution",
+        "build_graph_command_rejects_undeclared_host_effects",
+        "build_graph_command_multi_target_rejects_undeclared_host_effects",
+        "build_graph_command_rejects_graph_without_executable_targets",
+        "build_graph_command_accepts_library_dependencies",
+        "build_graph_command_rejects_gated_test_dependencies",
+        "build_graph_command_rejects_missing_root_source",
+        "build_program_lowering_collects_multiple_executable_targets",
+        "build_target_kind_owns_diagnostic_spelling",
+        "build_target_dsl_kind_owns_source_spelling",
+        "legacy_emit_json_modes_reject_build_zen_with_graph_diagnostic",
+        "emit_json_ast_marks_semantically_unchecked_sources_that_typed_json_rejects",
+        "emit_json_typed_command_outputs_checked_program",
+        "emit_json_diagnostics_command_outputs_machine_readable_errors",
+        "emit_json_diagnostics_includes_structured_return_keyword_fix",
+        "suggested_fixes",
+        "feature_gate",
+        "Type.implements(Behavior)",
+        "non-generic explicit behavior associations",
+        "collect_declarations_with_symbols_uses_resolver_behavior_impl_generic_method_template_target_and_name_metadata",
+        "collect_declarations_with_symbols_clears_stale_behavior_impl_generic_method_template_after_key_restore",
+        "resolver_declaration_metadata_skips_behavior_impl_methods_until_behavior_impl_pass",
+        "resolver_type_declaration_metadata_tasks_collect_only_type_work",
+        "resolver_callable_declaration_metadata_tasks_collect_callable_work",
+        "resolver_behavior_impl_block_declaration_tasks_collect_only_behavior_impls",
+        "resolver_type_reference_validation_tasks_collect_only_type_reference_work",
+        "declaration_collection_replay_bundle_collects_ast_and_resolver_tasks_together",
+        "resolver_declaration_semantic_bundle_replays_validation_passes",
+        "resolver_declaration_semantic_tasks_collect_only_semantic_work",
+        "resolver_declaration_semantic_bundle_replays_dedicated_semantic_tasks",
+        "resolver_declaration_collection_bundle_replays_metadata_semantics_and_refresh",
+        "generic_struct_constructor_without_type_args_is_error",
+        "nongeneric_struct_constructor_type_args_are_error",
+        "nongeneric_enum_constructor_type_args_are_error",
+        "nongeneric_struct_annotation_type_args_are_error",
+        "nongeneric_enum_annotation_type_args_are_error",
+        "behavior_impl_nongeneric_behavior_type_args_are_error",
+        "behavior_requires_nongeneric_behavior_type_args_are_error",
+        "behavior_extends_nongeneric_parent_type_args_are_error",
+        "generic_bound_nongeneric_behavior_type_args_are_error",
+        "generic_enum_constructor_without_type_args_is_error",
+        "async_scheduler_intrinsics_are_rejected_as_gated_not_unknown",
+        "effect_await_is_rejected_until_async_lowering_exists",
+        "typechecker_gated_methods_use_owned_action_enum",
+        "atomic_intrinsics_are_rejected_as_effect_gates",
+        "comptime_type_match_intrinsic_is_rejected_as_gated_not_unknown",
+        "typechecker_gated_intrinsics_use_owned_name_enum",
+        "dynamic_string_type_is_rejected_as_allocator_backed_gate",
+        "raw_memory_intrinsics_are_rejected_as_allocator_gates",
+        "byte_memory_intrinsics_are_rejected_as_allocator_gates",
+        "raw_pointer_intrinsics_are_rejected_as_ownership_gates",
+        "syscall_intrinsics_are_rejected_as_host_effect_gates",
+        "StaticString",
+        "allocator-backed `String`",
+        "type_declaration_keyword_owns_text_spelling",
+        "promoted_stdlib_modules_do_not_use_removed_or_gated_syntax",
+        "stdlib/io/net/unix_socket.zen",
+        "stdlib/io/net/socket.zen",
+        "stdlib/io/files/file.zen",
+        "stdlib/sys/process/prctl.zen",
+    ]
+    .into_iter()
+    .chain(
+        super::super::phase_audit_manifest::BUILD_DRIVER_PLAN_EVIDENCE
+            .iter()
+            .copied(),
+    ) {
+        assert!(
+            plan.contains(required),
+            "docs/PHASE_PLAN.md is missing durable plan text: {required}"
+        );
+    }
+
+    assert!(
+        !plan.contains("Generic behavior inheritance in `.extends` is still explicitly gated"),
+        "docs/PHASE_PLAN.md still claims generic behavior inheritance is gated"
+    );
+}
